requirements.txt
hashids~=1.3.1
Flask~=1.1.2
python-dotenv~=0.19.0
gunicorn==19.9.0
PyMySQL~=1.0.2
mysql-connector-python~=8.0.26